Backend/requirements.txt
Django==3.1.4
djangorestframework==3.13.1
typing_extensions==3.7.4.3
